# Conflicts:
#	routy-be/src/main/resources/mapper/mypage/mypageMapper.xml
#	routy-be/src/main/resources/mapper/plan/PlanMapper.xml

+++ routy-be/src/main/java/com/c4/routy/domain/region/dto/RegionDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class RegionDTO {
-    private Long regionId;
+    private Integer regionId;
     private String regionName;
     private String theme;
     private String regionDesc;
